// frontend/src/sections/prescriptions/list/createPrescription.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsertPrescriptionVariables = {
  storeId: string;
  input: {
    id: string;
    patientId: string;
    diagnosisId?: string | null;
    programId?: string | null;
    theirReference?: string | null;
    clinicianId?: string | null;
    prescriptionDate?: string | null;
  };
};

export type InsertPrescriptionResult = {
  insertPrescription: ({
  id: string;
});
};

export const InsertPrescription = {
  query: "mutation insertPrescription($storeId: String!, $input: InsertPrescriptionInput!) {\n  insertPrescription(storeId: $storeId, input: $input) {\n    ... on InvoiceNode {\n      id\n    }\n  }\n}",
} as TypedDocument<InsertPrescriptionResult, InsertPrescriptionVariables>;